% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commit_info.R
\name{get_commit_type}
\alias{get_commit_type}
\title{get commit type from user}
\usage{
get_commit_type()
}
\value{
a character vector
}
\description{
get commit type from user
}
\examples{
get_commit_type()
}
