% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_log_plot.R
\name{git_log_plot}
\alias{git_log_plot}
\title{git_log_plot}
\usage{
git_log_plot(git_log_df)
}
\arguments{
\item{git_log_df}{output of git_log() function}
}
\value{
ggplot2
}
\description{
git_log_plot
}
\examples{
git_log() \%>\% git_log_plot()
}
