% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitzen-package.R
\docType{package}
\name{gitzen-package}
\alias{gitzen}
\alias{gitzen-package}
\title{gitzen: Create Standard Git Commit Messages}
\description{
Commit messages are important means of communication between team members. This packages attempst to put some guidleines
  around how commit messages are created.
}
\author{
\strong{Maintainer}: Eric Lessert \email{elesert760@gmail.com}

}
\keyword{internal}
